Finalização dos endpoints de Cidade

+++ src/main/java/br/com/gocharge/repository/BandeiraRepository.java
@@ -1,35 +1,79 @@
 package br.com.gocharge.repository;
 
+import br.com.gocharge.domain.Bandeira;
+import br.com.gocharge.exceptions.NoContentException;
+import br.com.gocharge.exceptions.NotFoundException;
+import br.com.gocharge.mappers.BandeiraMapper;
 import br.com.gocharge.model.BandeiraModel;
+import br.com.gocharge.model.StatusCadastroModel;
 import org.springframework.stereotype.Repository;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
+import javax.transaction.Transactional;
 import java.util.List;
+import java.util.Objects;
 import java.util.UUID;
 
 @Repository
+@Transactional
 public class BandeiraRepository {
 
   @PersistenceContext EntityManager entityManager;
 
-  public List<BandeiraModel> getAll() {
-    return entityManager.createQuery("SELECT BandeiraModel FROM BandeiraModel").getResultList();
+  public List<Bandeira> getAll() {
+    List<BandeiraModel> bandeiras =
+        entityManager.createQuery("SELECT b FROM BandeiraModel b").getResultList();
+
+    if (bandeiras.size() > 0) {
+      return BandeiraMapper.INSTANCE.toDomain(bandeiras);
+    } else {
+      throw new NoContentException();
+    }
   }
 
-  public BandeiraModel getById(UUID id) {
-    return entityManager.find(BandeiraModel.class, id);
+  public Bandeira getById(UUID id) {
+    BandeiraModel bandeira = entityManager.find(BandeiraModel.class, id);
+
+    if (Objects.nonNull(bandeira)) {
+      return BandeiraMapper.INSTANCE.toDomain(bandeira);
+    } else {
+      throw new NotFoundException();
+    }
   }
 
-  public void create(BandeiraModel bandeira) {
-    entityManager.persist(bandeira);
+  public Bandeira create(Bandeira bandeira) {
+    BandeiraModel bandeiraModel = BandeiraMapper.INSTANCE.toModel(bandeira);
+
+    entityManager.persist(bandeiraModel);
+
+    return BandeiraMapper.INSTANCE.toDomain(bandeiraModel);
   }
 
-  public void update(BandeiraModel bandeira) {
-    entityManager.merge(bandeira);
+  public Bandeira update(Bandeira bandeira) {
+    BandeiraModel bandeiraModel = entityManager.find(BandeiraModel.class, bandeira.getId());
+
+    if (Objects.nonNull(bandeiraModel)) {
+      StatusCadastroModel statusCadastroModel = new StatusCadastroModel();
+      statusCadastroModel.setId(bandeira.getStatus());
+
+      bandeiraModel.setDescricao(bandeira.getDescricao());
+      bandeiraModel.setStatus(statusCadastroModel);
+
+      return BandeiraMapper.INSTANCE.toDomain(
+              entityManager.merge(bandeiraModel));
+    } else {
+      throw new NotFoundException();
+    }
   }
 
-  public void delete(BandeiraModel bandeira) {
-    entityManager.remove(bandeira);
+  public void delete(UUID idBandeira) {
+    BandeiraModel bandeira = entityManager.find(BandeiraModel.class, idBandeira);
+
+    if (Objects.nonNull(bandeira)) {
+      entityManager.remove(bandeira);
+    } else {
+      throw new NotFoundException();
+    }
   }
 }
